UPSTREAM: soc/intel/alderlake: Add method to determine the cpu type
set_cpu_type(): It determines the CPU type (big or small) that
is executing the function, and marks the global_cpu_type's array slot
which is corresponds to the executing CPU's index if the CPU type is
big core.

get_cpu_index(): It determines the index from LAPIC Ids. This is
required to expose CPPC3 package in ascending order of CPUs' LAPIC ids.
So, the function returns CPU's position from the ascending order list
of LAPIC ids.

TEST=Tested CPU index calculation, core type determination on Brya

Change-Id: Iaeb7491c178e574e633ca9048e4230831bce5b86
Signed-off-by: Nick Vaccaro <nvaccaro@google.com>
Original-Commit-Id: cd1cd8d1176cf65f189f0346679c3f4e4fe9af92
Original-Signed-off-by: Sridhar Siricilla <sridhar.siricilla@intel.com>
Original-Change-Id: If4ceb24d9bb1e808750bf618c29b2b9ea6d4191b
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/59360
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3403150
Reviewed-by: Kangheui Won <khwon@chromium.org>