UPSTREAM: device/pci_device.c: Improve pci_bridge_route() readability
Both the secondary and subordinate bus numbers are configured in this
function but it's not easy to search for in the tree as the PCI writes
are hidden inside a bigger write to 'PCI_PRIMARY_BUS'. Use separate
variables and PCI config writes to improve the readability.

(cherry picked from commit f879d36551dfa86128a691105dda376db50d0c3c)

Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/59131
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Change-Id: I3bafd6a2e1d3a0b8d1d43997868a787ce3940ca9
GitOrigin-RevId: f879d36551dfa86128a691105dda376db50d0c3c
Change-Id: Ib0dc5164cf83630a2807805d83a03be3b85d5cd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3500920
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>