UPSTREAM: utils: Add initial version of "remove_unused_code" script
This script creates a patch to remove all of the coreboot code that a
platform doesn't use.  This is useful for auditing the codebase for an
individual platform or releasing a platform's code.

Unlike the script that Sage used that did something similar, this keeps
the entire Kconfig tree (Though in a single file), all makefiles that
are required to build, and the standard build tools can still be used.
This will allow for much easier re-integration back into the coreboot
codebase if code is released after running this.

This is just the initial version and more features needed to be added to
make it fully functional.
- It should be able to build multiple configurations to retain the code
for all of those configurations.
- Flag to remove submodules files as well
- Additional variable flags to replace hardcoded values.
- The list of makefiles that need to be kept is pretty long, and could
be updated so that they aren't needed by the top level makefiles.
- Add flag to show changed files
- Show number of files before and after script is run

(cherry picked from commit 298b00776a139b9cea14319698bacf726b53da61)

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Change-Id: Iec69db2ad1358846d649db627b6d60ac8c2204e4
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/59169
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
GitOrigin-RevId: 298b00776a139b9cea14319698bacf726b53da61
Change-Id: I15bcec22d272a508b83fcfbefb7ec739f98133d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3823392
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>