mb/prodrive/hermes: Add VBT for Avalanche systems
The Hermes mainboard is used in different system configurations. The
current VBT for Poseidon systems is unsuitable for Avalanche systems
because display ports are connected differently.

Add a new field in the BMC config EEPROM layout and use it to choose
the correct VBT for every system configuration.

Change-Id: I2647f2ae3f496b9ad75980ba86beb7800fdb0668
Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/59838
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
GitOrigin-RevId: 4a0dee21ae97ea10eadae3bf5286ea6938b8eb89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3319063
Tested-by: Copybara Service <copybara-worker-blackhole@google.com>
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>