UPSTREAM: util/docker/coreboot.org-status: Rewrite parser
The current tool is a shell script that mixes data collection and HTML
generation and is generally a pain to work with. It takes 15 minutes to
run.

The new tool is written in go, collects all data first, then generates
the output HTML from the data and a single template, and finishes in
10 seconds.

The goal in this version is to produce output as similar as possible to
the output of the shell script. Some difference will remain because the
shell script returns some trash data whose reproduction would require
more effort than is worth.

(cherry picked from commit 3d0303a57c08bf8c56bb55885f6705680097bd27)

Original-Change-Id: I4fab86d24088e4f9eff434c21ce9caa077f3f9e2
Original-Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/59958
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Maxim Polyakov <max.senia.poliak@gmail.com>
GitOrigin-RevId: 3d0303a57c08bf8c56bb55885f6705680097bd27
Change-Id: I290200b12f3fff18f0bf21b7c7b21e508edb490f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4005975
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Patrick Georgi <pgeorgi@google.com>
Reviewed-by: Kangheui Won <khwon@chromium.org>